# components/WebServer/public/apidoc/swagger.yml
openapi: '3.0.0'
info:
  version: '1.0.0'
  title: 'Conversation manager API client'
  description: 'a tool for listening to, reading, editing and annotating a conversation transcript'
  termsOfService: https://doc.linto.ai/#/?id=gnu-affero-general-public-licence-v3
  contact:
    name: LINAGORA
    url: http://linto.ai
    email: contact@linto.ai
#servers:
host: localhost:8000
basePath: /
schemes:
 - https
paths:
...
  /login: 
    post:
      tags:
        - "users"
      summary: Logs in a user and returns the authentication cookie
      requestBody:
        required: true
        description: A JSON object containing the username and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      # security: []    # no authentication
      responses:
        '200':
          description: 
            Successfully authenticated.
            The session ID is returned in a cookie named `JSESSIONID`. You need to include this cookie in subsequent requests.
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: JSESSIONID=abcde12345; Path=/; HttpOnly
        '400':
          description: 
        '404':
          description:
        default:
          description:
  /api/user: 
    post:
      tags:
        - "users"
      operationId: createUser
      summary: creates a user
      requestBody:
        required: true
        description: A JSON object containing the username and password.
        content:   
          application/json:
            schema: 
              type: object
              properties:
                userName: 
                  type: string
                password:
                  type: string
                email:
                  type: string
                  format: email
                # convoAccess:
                #   type: array
                #   items: 
                #     type: object 
                #     properties:
                #       conversationid:
                #         type: string 
                #       accessType:
                #         type: string
      responses:
        '201':
          description: user is successfully created
          schema:
            $ref: '#/components/schemas/User'
        '400':
          description: bad request error
          schema:
            $ref: '#components/schemas/Error'
  /api/user/{userid}:
    get: 
      tags:
        - "users"
      operationId: getUser
      summary: Gets a user by id
      description: gets a single user
      parameters:
        - in: path
          name: userid
          schema:
            type: string
          required: true
      responses:
        '200':
          description: json document 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: bad request error
          schema:
            $ref: '#components/schemas/Error'
    put:
      tags:
        - "users"
      summary: updates a user password
      description: Adds conversation(s) id to a user's accessedMeetings
      requestBody:
        required: true 
        description: A JSON object containing the user id and conversation id(s).
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                accessedMeetings:
                  type: array
                  items:
                    type: integer 
      responses:
        '200':
          description: Successfully updated.
        '400': 
          description: Invalid id supplied 
        '404':
          description: User not found 
    delete:
      tags:
        - "users"
      summary: deletes a user by id
      parameters:
        - in: path
          name: userid
          schema:
            type: string
          required: true
      responses:
        '200':
          description: json document 
        '400':
          description: bad request error
          schema:
            $ref: '#components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID

  # parameters:
  # requestBodies:
  # responses:
  # headers:
  # examples:
  # links:
  # callbacks:
  schemas:
...
    UserLogin:
      type: object
      properties:
        userName:
          type: string 
        password:
          type: string
          format: password 
          writeOnly: true
      required:
        - userName
        - password
    User:
      type: object
      properties:
        userid:
          type: string
          readOnly: true
        userName:
          type: string 
        email:
          type: string
          format: email
        convoAccess:
          type: array
          items: 
            type: object 
            properties:
              conversationid:
                type: string 
              accessType:
                type: string
        required:
          - userid
          - userName
          - convoAcess 
...
security:
  - cookieAuth: []